fixup! function moving and varaible renaming
